worker/requirements.txt
SQLAlchemy==2.0.12
psycopg2==2.9.6
PyYAML==6.0
httpx[http2,brotli]==0.24.1
python_dotenv==1.0.0
tweepy==4.14.0
praw==7.7.0
Mastodon.py==1.8.1
schedule==1.2.0